requirements.txt
Django==5.2.7
djangorestframework==3.15.2
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.0
Pillow==11.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
reportlab==4.2.5
django-cors-headers==4.6.0
dj-database-url==2.1.0
drf-spectacular==0.27.2

